+++ src/core/resizePage.ts
@@ -1,22 +1,38 @@
-import * as constants from '~/constants';
 import createPageNode from '~/core/createPageNode';
+
+// The sequence number of the drawing.
+// This ensures that if multiple resizes are requested at the same time, the page of the latest resize request is drawn.
+let draw = -1;
 
 /**
   * Resize page.
   */
-export default (pages: any[], zoomFactor: number = 1): void => {
-  for (let pageNumber=1; pageNumber<=pages.length; pageNumber++) {
-    // Fetch page.
-    const page = pages[pageNumber - 1];
+export default async (pages: any[], zoomFactor: number = 1): Promise<void> => {
+  (async (pages: any[], zoomFactor: number, currentDraw: number) => {
+    // console.log(`Start resizing(${currentDraw})`);
+    for (let pageNumber=1, numberOfPages=pages.length; pageNumber<=numberOfPages; pageNumber++) {
+      if (draw !== currentDraw) {
+        // If it is not the latest resize request, the drawing is canceled.
+        // console.log(`Cancel resize(${currentDraw})`);
+        return;
+      }
 
-    // Create a page node.
-    const [_, canvas, viewport, devicePixelRatio] = createPageNode('resize', page, pageNumber, zoomFactor);
+      // Fetch page.
+      const page = pages[pageNumber - 1];
 
-    // Render page content on canvas.
-    page.render({
-      canvasContext: canvas.getContext('2d'), 
-      transform: devicePixelRatio !== 1 ? [devicePixelRatio, 0, 0, devicePixelRatio, 0, 0] : null,
-      viewport
-    });
-  }
+      // Create a page node.
+      const [_, canvas, viewport, devicePixelRatio] = createPageNode('resize', page, pageNumber, zoomFactor);
+
+      // Render page content on canvas.
+      const task = page.render({
+        canvasContext: canvas.getContext('2d'), 
+        transform: devicePixelRatio !== 1 ? [devicePixelRatio, 0, 0, devicePixelRatio, 0, 0] : null,
+        viewport
+      });
+
+      // Wait for rendering to complete.
+      await task.promise;
+      // console.log(`Resize page ${pageNumber} of ${numberOfPages}(${currentDraw})`);
+    }
+  })(pages, zoomFactor, ++draw);
 }