+++ types/constants.d.ts
@@ -2,3 +2,5 @@ export declare const MINIMUM_MOBILE_WIDTH: number;
 export declare const ZOOM_OUT: number;
 export declare const ZOOM_IN: number;
 export declare const PDF_DRAWING_SCALE: number;
+export declare const RESIZE_DELAY_SECONDS: number;
+export declare const CANVAS_AREA_LIMIT: number;