+++ types/core/resizePage.d.ts
@@ -1,4 +1,4 @@
-declare const _default: (pages: any[], zoomFactor?: number) => void;
+declare const _default: (pages: any[], zoomFactor?: number) => Promise<void>;
 /**
   * Resize page.
   */